backend/currency_api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.108.0
flake8==7.1.0
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.32.0.post1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.25.0
uvloop==0.19.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.13.0
yarl==1.9.4
